Move proof.go file to a new package
