+++ log_proofs.go
@@ -18,15 +18,12 @@ import (
 	"errors"
 	"fmt"
 
-	"github.com/transparency-dev/merkle/compact"
+	"github.com/transparency-dev/merkle/proof"
 )
 
 // NodeFetch bundles a node ID with additional information on how to use the
 // node to construct a proof.
-type NodeFetch struct {
-	ID     compact.NodeID
-	Rehash bool
-}
+type NodeFetch = proof.NodeFetch
 
 // CalcInclusionProofNodeAddresses returns the tree node IDs needed to build an
 // inclusion proof for a specified tree size and leaf index. All the returned
@@ -40,7 +37,7 @@ func CalcInclusionProofNodeAddresses(size, index uint64) ([]NodeFetch, error) {
 	if index >= size {
 		return nil, fmt.Errorf("invalid parameter for inclusion proof: index %d is >= size %d", index, size)
 	}
-	return proofNodes(index, 0, size, true), nil
+	return proof.Nodes(index, 0, size, true), nil
 }
 
 // CalcConsistencyProofNodeAddresses returns the tree node IDs needed to build
@@ -59,7 +56,7 @@ func CalcConsistencyProofNodeAddresses(size1, size2 uint64) ([]NodeFetch, error)
 		return nil, fmt.Errorf("invalid parameter for consistency proof: size1 %d > size2 %d", size1, size2)
 	}
 
-	return consistencyNodes(size1, size2), nil
+	return proof.Consistency(size1, size2), nil
 }
 
 // Rehash computes the proof based on the slice of NodeFetch structs, and the
